+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- feat: Read Carbon version from package.json
+- chore: Change `target` option to `testOnlyTarget`
+
 ## 2.0.0-beta.1
 
 - Restore optional v10 support.